Refactor Google Services
- Bersihin MokuApp.swift
- Nambahin Main View / Entry Point buat Bengkel & Customer
