+++ Moku/MokuApp.swift
@@ -7,156 +7,32 @@
 
 import SwiftUI
 import Firebase
-import GoogleMaps
-import GooglePlaces
-import Combine
-
-struct BengkelView: View {
-    @State var bengkel: Bengkel
-
-    init(from bengkel: Bengkel) {
-        _bengkel = State(wrappedValue: bengkel)
-    }
-
-    var body: some View {
-        Text("Bengkel View")
-    }
-}
-
-struct CustomerView: View {
-    @State var customer: Customer
-
-    init(from customer: Customer) {
-        _customer = State(wrappedValue: customer)
-    }
-
-    var body: some View {
-        Text("\(customer.name) - \(customer.phoneNumber)")
-    }
-}
 
 @main
 struct MokuApp: App {
-    private lazy var locationManager: CLLocationManager = {
-        let manager = CLLocationManager()
-        manager.delegate = LocationManager.shared
-        manager.desiredAccuracy = kCLLocationAccuracyBest
-        return manager
-    }()
-
-    @ObservedObject var userLocation = LocationManager.shared
     @ObservedObject var session = SessionService.shared
-
-    var subsriptions = Set<AnyCancellable>()
 
     init() {
         FirebaseApp.configure()
-        GMSServices.provideAPIKey("AIzaSyDs--hrfb86N3WtmfbMXMoah4MiZcfzLF4")
-        GMSPlacesClient.provideAPIKey("AIzaSyDs--hrfb86N3WtmfbMXMoah4MiZcfzLF4")
 
-        locationManager.requestWhenInUseAuthorization()
+        LocationService.shared.requestUserAuthorization()
 
-        GooglePlacesService.shared.$results.sink { places in
-            places.forEach { place in
-                print(place.name)
-            }
-        }.store(in: &subsriptions)
-
-        DispatchQueue.main.asyncAfter(deadline: .now() + 3) {
-            GooglePlacesService.shared.runQuery(" bangka  ")
-        }
+        GooglePlacesService.register()
+        GoogleMapsService.register()
     }
 
     var body: some Scene {
         WindowGroup {
-            //            ContentView()
             if let user = session.user {
                 switch user {
                 case let .bengkel(bengkel):
-                    BengkelView(from: bengkel)
+                    BengkelView(for: bengkel)
                 case let .customer(customer):
-                    CustomerView(from: customer)
+                    CustomerView(for: customer)
                 }
             } else {
-                // Suruh Login...
-                GoogleMapView(coordinate: $userLocation.coordinate) { coordinate in
-                    MapHelper.geocode(coordinate: coordinate) { location in
-                        print(location)
-                    }
-                }
+                // ...
             }
         }
-    }
-}
-
-final class LocationManager: NSObject, CLLocationManagerDelegate, ObservableObject {
-    @Published var coordinate = CLLocationCoordinate2D(latitude: 5, longitude: 105)
-
-    static let shared = LocationManager()
-
-    func locationManagerDidChangeAuthorization(_ manager: CLLocationManager) {
-        if manager.authorizationStatus == .authorizedWhenInUse {
-            manager.requestLocation()
-        }
-    }
-
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        if let recentLocation = locations.last {
-            coordinate = recentLocation.coordinate
-        }
-    }
-
-    func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
-        print(error.localizedDescription)
-    }
-}
-
-struct GoogleMapView: UIViewRepresentable {
-    @Binding var coordinate: CLLocationCoordinate2D
-
-    private let mapView = GMSMapView(frame: .zero)
-    private let zoomLevel: Float = 18
-
-    let onAnimationEnded: (CLLocationCoordinate2D) -> Void
-
-    func makeUIView(context: Context) -> GMSMapView {
-        mapView.camera = GMSCameraPosition.camera(withTarget: coordinate, zoom: zoomLevel)
-        mapView.delegate = context.coordinator
-        mapView.isUserInteractionEnabled = true
-
-        marker.map = mapView
-
-        return mapView
-    }
-
-    func updateUIView(_ uiView: GMSMapView, context: Context) {
-        uiView.animate(toLocation: coordinate)
-        uiView.clear()
-        marker.map = uiView
-    }
-
-    func makeCoordinator() -> GoogleMapViewCoordinator {
-        GoogleMapViewCoordinator(self)
-    }
-
-    final class GoogleMapViewCoordinator: NSObject, GMSMapViewDelegate {
-        var mapView: GoogleMapView
-
-        init(_ mapView: GoogleMapView) {
-            self.mapView = mapView
-        }
-
-        func mapView(_ mapView: GMSMapView, didEndDragging marker: GMSMarker) {
-            self.mapView.onAnimationEnded(marker.position)
-        }
-    }
-}
-
-extension GoogleMapView {
-    private var marker: GMSMarker {
-        let marker = GMSMarker(position: coordinate)
-        marker.isDraggable = true
-        marker.appearAnimation = .pop
-        return marker
     }
 }
